+++ example/lib/examples/delete_confirmation.dart
@@ -1,5 +1,3 @@
-import 'dart:math';
-
 import 'package:cue/cue.dart';
 import 'package:flutter/material.dart';
 import 'package:iconsax/iconsax.dart';