+++ example/lib/main.dart
@@ -1,4 +1,5 @@
 import 'package:cue/cue.dart';
+import 'package:example/examples/x_style_side_menu.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
@@ -24,8 +25,8 @@ class CueApp extends StatelessWidget {
           seedColor: const Color(0xFF6C63FF),
         ),
       ).copyWith(splashFactory: InkSparkle.splashFactory),
-      themeMode: ThemeMode.light,
-      home: BasiceExample(),
+      themeMode: ThemeMode.system,
+      home: XStyleSideMenu(),
       showPerformanceOverlay: false,
       debugShowCheckedModeBanner: false,
       builder: (context, child) {
@@ -66,7 +67,7 @@ class _BasiceExampleState extends State<BasiceExample> {
               ],
               child: CueDragScrubber(
                 distance: 100,
-                scrubForwardDirection: CueScrubAxis.up,
+                scrubForwardDirection: .up,
                 releaseMode: .reverse,
                 child: Container(
                   width: 100,