+++ test/acts/card_test.dart
@@ -522,7 +522,9 @@ void main() {
 
     testWidgets('apply with hasBorderStroke and borderOnForeground false uses painter', (tester) async {
       const act = CardAct(
-        shape: AnimatableValue.fixed(BeveledRectangleBorder()),
+        shape: AnimatableValue.fixed(BeveledRectangleBorder(
+           side: BorderSide(color: Colors.red, width: 2),
+        )),
         borderOnForeground: false,
       );